LPS-107538 portal-search: Extract the anonymous instatiation of ModelIndexerWriterDocumentHelper to a named class we can test.
https://issues.liferay.com/browse/LPS-107538
