+++ modules/apps/portal-search/portal-search/src/main/resources/META-INF/module-log4j.xml
@@ -6,6 +6,10 @@
 		<priority value="WARN" />
 	</category>
 
+	<category name="com.liferay.portal.search.internal.index.contributor.helper.ModelIndexerWriterDocumentHelperImpl">
+		<priority value="WARN" />
+	</category>
+
 	<category name="com.liferay.portal.search.internal.indexer.IndexerWriterImpl">
 		<priority value="WARN" />
 	</category>
